Merged PR 25390: Fix crash when resizing an HWND CanvasSwapChain
This PR fixes an issue I've found while putting together a Win32 WASDK sample using Win2D.

The issue is that `CanvasSwapChain` assumes the underlying `IDXGISwapChain1` in use is either one for `CoreWindow`, or one for Composition. So, if you create a `CanvasSwapChain` through the interop `GetOrCreate` helper by wrapping a HWND swapchain, and then try to resize it, Win2D will incorrectly think it can set the transform matrix for the swapchain (as it sees the swapchain doesn't have a `CoreWindow`, so it assumes it must be a Composition swapchain), which then crashes. This makes it so that you can't resize a `CanvasSwapChain` wrapping a HWND-based `IDXGISwapChain1` object created through manual interop.

The fix is pretty simple: it adds a check for HWND swapchains as well, and if that's the case it also skips setting the transform matrix when resizing the swap chain, just like Win2D already does for `CoreWindow`.

No public API changes, just an additional check only done when creating a wrapping swapchain.
That is, this extra check is _only_ done in interop scenarios, and not when creating the wrapper internally.