Merged PR 29945: Roll Target Framework Version back to 19041
Win2D updated its TargetPlatformVersion and TargetFrameworkVersion to 22621 ahead of WinAppSDK, which is still at 19041. This meant that any apps using the WinUI3 Visual Studio templates would not build out of the box when adding the Win2D nuget package.

This change rolls back to 19041, which we expect will not break any scenarios for existing customers who have already updated to 22621.