Merged PR 31430: Update WinAppSDK to 1.6.240531000-experimental1
Update WinAppSDK to 1.6.240531000-experimental1 and CsWinRT to 2.1.0-prerelease.240602.1

This change updates TargetPlatformVersion and TargetFramework version from 19041 to 22000. In WinAppSDK, the experimental channel uses 22000 instead of 19041, for reasons related to stripping out experimental APIs. When WinAppSDK 1.6 stable releases, it will drop back to 19041. Win2D will also drop back down to 19041 for the stable release.