Merged PR 27700: Reenable CanvasAnimatedControl for WinUI3
Reenable CanvasAnimatedControl for WinUI3.

This mostly involved changing CoreDispatcher over to DispatcherQueue. This has some caveats as the game thread was relying on receiving the Completed event for the IAsyncAction that is returned by CoreDispatcher.RunAsync, but the equivalent DispatcherQueue.TryEnqueue returns void. To work around this, we wrap the callback in an IAsyncAction, which then gets manually called and fires a completion event as part of the Callback that is passed to the DispatcherQueue.

This change also reenables the CanvasAnimatedControl tests (minus a few that did not survive the port to WinUI3) and removes the CANVAS_ANIMATED_CONTROL_IS_ENABLED macro.