Merged PR 31886: Add .NET 8 projections for Win2D.uwp
### Overview

This PR adds a .NET projections project for Win2D.uwp. It's basically identical to the one in the WinAppSDK branch, with the only difference being that this doesn't actually use WinAppSDK, but Windows.UI.Xaml.

I also backported the fix for the RIDs for the native libraries, switching `win10-` to just `win-`.

### Work items

https://microsoft.visualstudio.com/OS/_workitems/edit/51723527