Merged PR 32595: 🍒⛏️ 24714: Strong sign .NET assembly with public key
This PR enables strong name assembly signing for the .NET projections.

![image.png](https://mscodehub.visualstudio.com/8b1f3307-e83a-4982-b590-048338090061/_apis/git/repositories/dc534cf5-25c9-4f57-83b0-7a52fa2a046d/pullRequests/24714/attachments/image.png)

Strong name signing is **not** a security boundary, so keeping the key in the repo is fine. For context, we've been doing the same for years in the Toolkit repos as well. The [docs](https://learn.microsoft.com/en-us/dotnet/standard/library-guidance/strong-naming) also say it's fine to keep the key public.

Cherry picked from !24714

Cherry-picked from commit `a9492ece`.

----
#### AI description  (iteration 1)
#### PR Classification
Code signing enhancement.

#### PR Summary
This pull request adds strong naming to the .NET assembly to ensure its authenticity and integrity.
- `build/Win2D.cs.targets`: Enabled strong assembly signing and specified the key file for signing.