Merged PR 33546: Fix issue where pausing a CAC a second time results in 100% CPU usage on thread
Fix issue where pausing a CAC a second time results in 100% CPU usage on thread. The issue was that calling DispatcherQueue.RunEventLoop was expected to block until someone calls DispatcherQueue.EnqueueEventLoopExit. This happens for the first time, however after EnqueueEventLoopExit has been called, another call of RunEventLoop does not block until the method is called a second time but rather just finished the queue and returns. This resulted in 100% CPU usage on the GameLoopThread since game loop was continously calling RunEventLoop since it wasnt blocking.

By using the the RunEventLoopWithOptions overload with DispatcherRunOptions::DispatcherRunOptions_QuitOnlyLocalLoop, we achieve the desired result.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses an issue where pausing a CAC a second time results in 100% CPU usage on a thread.
- `winrt/lib/xaml/GameLoopThread.cpp`: Modified the event loop call to use `RunEventLoopWithOptions` with `DispatcherRunOptions_QuitOnlyLocalLoop` to prevent excessive CPU usage.